requirements.txt
colorama==0.4.6
logit-axis==0.1.5
pygame-ce==2.1.4
